Added/updated admin analytics/refund, customer dispute/edit/profile/notification screens and profile repository

+++ lib/presentation/screens/admin/admin_analytics_dashboard.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:servyn/services/analytics_service.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-import '../../services/analytics_service.dart';
 
 class AdminAnalyticsDashboard extends StatefulWidget {
   const AdminAnalyticsDashboard();
@@ -239,7 +239,7 @@ class _AdminAnalyticsDashboardState extends State<AdminAnalyticsDashboard> {
                               ClipRRect(
                                 borderRadius: BorderRadius.circular(4),
                                 child: LinearProgressIndicator(
-                                  value: percentage,
+                                  value: percentage.toDouble(),
                                   minHeight: 8,
                                 ),
                               ),
